+++ wp-content/plugins/wordpress-beta-tester/CHANGES.md
@@ -1,5 +1,9 @@
 [unreleased]
 
+#### 2.2.12 / 2020-08-10
+* fix intermittent PHP warning [#21](https://github.com/afragen/wordpress-beta-tester/issues/21)
+* deactivate and die if user attempting to run with `wordpress-develop`
+
 #### 2.2.11 / 2020-08-01
 * minor cleanup